sentinel-core/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sentinel-core"
version = "0.50.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.44.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.13.0",
    "httpx>=0.28.1",
    "tiktoken",
    # AI provider libraries — pinned per supply-chain incident (LiteLLM 1.82.7-1.82.8 were malicious, March 2026)
    "litellm>=1.83.0,<2.0",
    "tenacity>=8.2.0,<10.0",
    "anthropic>=0.93.0,<1.0",
    "numpy>=1.26,<3.0",
    "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
pythonpath = ["../shared"]

[tool.ruff]
line-length = 100
target-version = "py312"
